bump versions, update AD tests, update workflows (#166)
This fixes all the existing CompatHelper PRs

Closes #160
Closes #161
Closes #163 
Closes #164 

And also updates the AD testing to use Mooncake (and removes the
previously commented-out Tracker and Zygote tests).

And also updates the test and format workflows to more closely match the
rest of TuringLang. In particular:
- uses
[TuringLang/actions/Format](https://github.com/TuringLang/actions?tab=readme-ov-file#format)
for workflow standardisation
- tests on both Julia 1.10 and Julia 1.11
- removes the `arch=x64` matrix variable which is no longer appropriate
for macOS-latest (should be `aarch64`, but leaving it out is easier as
it will allow GitHub to automatically determine the right architecture)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>